requirements.txt
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
